--- src/AutoHeader/docs/AutoHeader.md
@@ -1,3 +1,0 @@
-# AutoHeader
-
-TODO
